refactor: constants into utils file
